+++ Sound/SC3/Server/State/Monad/Process.hs
@@ -1,4 +1,4 @@
-module Sound.SC3.Server.Monad.Process (
+module Sound.SC3.Server.State.Monad.Process (
     withSynth
   , withDefaultSynth
   -- * Re-exported for convenience
@@ -7,10 +7,10 @@ module Sound.SC3.Server.Monad.Process (
 
 import           Data.Default (def)
 import qualified Sound.SC3.Server.Connection as Conn
-import           Sound.SC3.Server.Monad (Server)
-import qualified Sound.SC3.Server.Monad as Server
 import           Sound.SC3.Server.Process hiding (withSynth)
 import qualified Sound.SC3.Server.Process as Process
+import           Sound.SC3.Server.State.Monad (Server)
+import qualified Sound.SC3.Server.State.Monad as Server
 
 withSynth :: ServerOptions -> RTOptions -> OutputHandler -> Server a -> IO a
 withSynth serverOptions rtOptions outputHandler action =