+++ examples/sine-grains.hs
@@ -2,12 +2,12 @@ import           Control.Concurrent.MVar
 import           Control.Monad (void, when)
 import           Control.Monad.IO.Class (MonadIO, liftIO)
 import           Sound.SC3.UGen
-import           Sound.SC3.Server.Monad
-import           Sound.SC3.Server.Monad.Command
+import           Sound.SC3.Server.Notification
+import           Sound.SC3.Server.State.Monad
+import           Sound.SC3.Server.State.Monad.Command
 -- You need the hsc3-server-internal package in order to use the internal server
 --import           Sound.SC3.Server.Monad.Process.Internal (withDefaultInternal)
-import           Sound.SC3.Server.Monad.Process (withDefaultSynth)
-import           Sound.SC3.Server.Notification
+import           Sound.SC3.Server.State.Monad.Process (withDefaultSynth)
 import           Sound.OpenSoundControl (immediately)
 import qualified Sound.OpenSoundControl as OSC
 import           System.Posix.Signals (installHandler, keyboardSignal, Handler(Catch))