+++ ChangeLog.md
@@ -1,3 +1,12 @@
+### v0.5.0
+
+* Rename `Sound.SC3.Server.Monad.Send` to `Sound.SC3.Server.Monad.Request` and refactor interface and semantics
+* Use distinct types for audio and control buses and IDs
+* Factor monad type classes into `Sound.SC3.Server.Monad.Class`
+* Remove `Sound.SC3.Server.Monad.capture` combinator
+* Upgrade to `hosc` 0.13 and `hsc3` 0.13
+* Rename `Sound.SC3.Server.Monad*` to `Sound.SC3.Server.State.Monad*`
+
 ### v0.4.0
 
 * Move allocation interface to `Sound.SC3.Server.Monad` and add monad control
@@ -13,4 +22,3 @@
 ### v0.3.0
 
 * Add new modules `Sound.SC3.Server.Monad.Send` and `Sound.SC3.Server.Monad.Command` providing abstractions for synchronous and asynchronous commands and completion messages
-